Fix bug in repeated collapse of empty bodies
The initial rendering of the empty body wrote '' to the format cache,
and observes it. The second rendering reads from the cache, but then
because the value is falsey it writes to the cache again! This write
side-effect in a computed on an observable is bad bad bad and breaks the
rendering.

We now properly distinguish cache presence, not just falsey-ness.